Change sprite layering order.
Player sprites are in front of PickUpObjects. The carried PickUpObject
is in front of the Player. Players and PickUpObjects are in front of
Boxes.

To accomplish this, put everything in one layer and then change the
sorting order.

Also changed the Player prefab default sprite to something other than
the rectangle.